[refactor] component files
