+++ renderer/components/items/hearts.tsx
